cli/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ct"
version = "0.1.0"
description = "Unified CLI for ComfyUI development tools"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    # CLI framework
    "typer[all]>=0.9.0",
    "rich>=13.0.0",

    # GitHub API
    "PyGithub>=2.1.1",

    # Reddit API
    "praw>=7.7.1",

    # Web dashboard
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "jinja2>=3.1.2",
    "python-multipart>=0.0.6",

    # Utilities
    "python-dotenv>=1.0.0",
    "httpx>=0.26.0",
    "tqdm>=4.66.0",
    "pyyaml>=6.0.0",

    # Discord bot
    "discord.py>=2.3.0",

    # Google Sheets
    "gspread>=5.12.0",
    "google-auth>=2.0.0",
]

[project.optional-dependencies]
pdf = ["marker-pdf>=1.0.0"]

[project.scripts]
cds = "main:app"

[tool.setuptools]
py-modules = ["main", "config"]

[tool.setuptools.packages.find]
where = ["."]
